+++ packages/components/ui/src/bin/makeFeatherIcons.js
@@ -4,7 +4,13 @@ import path from 'path';
 
 import * as _ from 'lamb';
 import {tapMessage} from '@svizzle/dev';
-import {readDirFilesIndexed, saveString, writeFile} from '@svizzle/file';
+import {
+	readDirFilesIndexed,
+	resolveToDir,
+	saveObj,
+	saveString,
+	writeFile,
+} from '@svizzle/file';
 import {
 	makePostfixed,
 	makePrefixed,
@@ -17,14 +23,16 @@ import {parse as parseSVG} from 'svg-parser';
 // Run after `npm run lernacleanboot` at global level.
 // Lerna installs all deps in `<root of the repo>/node_modules/`
 const FEATHER_ICONS_DIR =
-	path.resolve(__dirname, '../../../../node_modules/feather-icons/dist/icons');
-const OUTPUT_DIR = path.resolve(__dirname, '../src/icons/feather');
+	path.resolve(__dirname, '../../../../../node_modules/feather-icons/dist/icons');
+const ROLLUP_INPUT_PATH = path.resolve(__dirname, '../rollup/rollupIconsInput.json');
+const FEATHER_GLYPHS_DIR = path.resolve(__dirname, '../icons/feather');
 
-export const isSvgFile = filepath => path.parse(filepath).ext.endsWith('.svg');
-export const isSvgTag = _.pipe([_.getKey('tagName'), _.is('svg')]);
-export const stringifySvgTags = str => toHtml(str, {space: 'svg'});
-export const camelCase = str => camelcase(str, {pascalCase: true});
-export const extractSvgBody = _.pipe([
+const resolveToFeatherDir = resolveToDir(FEATHER_GLYPHS_DIR);
+const isSvgFile = filepath => path.parse(filepath).ext.endsWith('.svg');
+const isSvgTag = _.pipe([_.getKey('tagName'), _.is('svg')]);
+const stringifySvgTags = str => toHtml(str, {space: 'svg'});
+const camelCase = str => camelcase(str, {pascalCase: true});
+const extractSvgBody = _.pipe([
 	parseSVG,
 	_.getKey('children'),
 	_.findWhere(isSvgTag),
@@ -34,39 +42,47 @@ export const extractSvgBody = _.pipe([
 	makePrefixed(`<svelte:options namespace='svg'/>\n`),
 	makePostfixed('\n'),
 ]);
-export const resolveTo = dirPath => filename => path.resolve(dirPath, filename);
-export const resolve = resolveTo(OUTPUT_DIR);
-export const makeComponentName = _.pipe([
+const makeComponentName = _.pipe([
 	splitByDot,
 	_.getAt(0),
 	camelCase,
 ]);
-export const makeComponentPath = _.pipe([
+const makeComponentPath = _.pipe([
 	makeComponentName,
 	makePostfixed('.svelte'),
-	resolve
+	resolveToFeatherDir
 ]);
-export const writeMapToFiles = obj => _.map(
+const writeMapToFiles = obj => _.map(
 	_.pairs(obj),
-	([filepath, string]) =>
-		writeFile(makeComponentPath(filepath), string, 'utf8')
+	([filename, string]) =>
+		writeFile(makeComponentPath(filename), string, 'utf8')
 );
-
-export const writeIndex = _.pipe([
+const writeFeatherIndex = _.pipe([
 	_.keys,
 	_.mapWith(_.pipe([
 		makeComponentName,
 		id => `export {default as ${id}} from './${id}.svelte';`
 	])),
 	_.joinWith('\n'),
-	saveString(resolve('index.js'))
+	saveString(resolveToFeatherDir('index.js'))
 ]);
-export const createFeatherIcons = obj => Promise.all([
+const createRollupIconsInput = _.pipe([
+	_.keys,
+	_.mapWith(_.collect([
+		makeComponentName,
+		id => `src/icons/feather/${makeComponentName(id)}.svelte`
+	])),
+	_.fromPairs,
+	saveObj(ROLLUP_INPUT_PATH, 2)
+]);
+const setupFeatherIcons = obj => Promise.all([
 	...writeMapToFiles(obj),
-	writeIndex(obj),
+	writeFeatherIndex(obj),
+	createRollupIconsInput(obj),
 ]);
 
 readDirFilesIndexed(FEATHER_ICONS_DIR, isSvgFile, extractSvgBody)
-.then(createFeatherIcons)
-.then(tapMessage(`Saved files in ${OUTPUT_DIR}`))
+.then(setupFeatherIcons)
+.then(tapMessage(`Saved files in ${ROLLUP_INPUT_PATH}`))
+.then(tapMessage(`Saved files in ${FEATHER_GLYPHS_DIR}`))
 .catch(err => console.error(err));