+++ packages/tools/dev/src/index.js
@@ -1,5 +1,0 @@
-export * from './log/array-[any-any]'
-export * from './log/iterable-array'
-export * from './log/string-[any-any]'
-export * from './common'
-export * from './test'