+++ packages/tools/file/src/path.spec.js
@@ -3,7 +3,8 @@ import {strict as assert} from 'assert';
 import {
 	filterJsonExtensions,
 	hasAnyExtensionOf,
-	renameToExtension
+	renameToExtension,
+	resolveToDir,
 } from './path';
 
 describe('path', function() {
@@ -48,6 +49,20 @@ describe('path', function() {
 			assert.deepStrictEqual(
 				renameToMinJs('file.txt'),
 				'file.min.js'
+			);
+		});
+	});
+	describe('resolveToDir', function() {
+		it('should return a function expecting a filename and returning it resolved to the provided directory path', function() {
+			const resolve = resolveToDir('/output/dir/');
+
+			assert.deepStrictEqual(
+				resolve('file1.txt'),
+				'/output/dir/file1.txt'
+			);
+			assert.deepStrictEqual(
+				resolve('file2.txt'),
+				'/output/dir/file2.txt'
 			);
 		});
 	});